Fixed bug with HTML parsing
Fixed bug with the parsing of html. Resulting buffer after html is parsed was computed based on the remaining buffered output after running the tokenizer, but needs to be computed based on the remaining buffer and the remaining reader reader data.

This resulted in inaccurate Go output pre-ast computation, which pretty much broke everything.